refact: BlockProvider has been moved to provider package
