+++ pathetic-api/src/main/java/de/metaphoriker/pathetic/api/pathing/filter/filters/WaterPathFilter.java
@@ -4,7 +4,7 @@ import lombok.NonNull;
 import org.bukkit.Material;
 import de.metaphoriker.pathetic.api.pathing.filter.PathValidationContext;
 import de.metaphoriker.pathetic.api.pathing.filter.PathFilter;
-import de.metaphoriker.pathetic.api.snapshot.BlockProvider;
+import de.metaphoriker.pathetic.api.provider.BlockProvider;
 import de.metaphoriker.pathetic.api.wrapper.PathPosition;
 
 /** A PathFilter implementation that determines if a path is through water. */
